profiler_requirements.txt
pandas
memory_profiler
Click
PyYAML